XF3-827 : DPoE details are not displayed on Gateway>Connection>Xfinity Network
Reason for change: Changed codes to return all link entries regardless of the
forwarding state.  This is needed because Ccsp only ask for this once as soon
as EPON agent is ready while the DPoE may not be ready yet.  This causes all
16 entries to be returned and the UI will display all 16 entries.

Test Procedure: DPoE details should be filled in.

Change-Id: I4a5f0be6fea28ce38fe4d90ea2e27e3c6335ee2a
Risks: None
Signed-off-by: yfong <Yick.Fong@pace.com>